requirements.txt
flask
oursql
pyaml
